% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{accumulate_threshold}
\alias{accumulate_threshold}
\title{Accumulate vector based on conditional threshold}
\usage{
accumulate_threshold(x, threshold)
}
\arguments{
\item{x}{Numeric vector.}

\item{threshold}{Integer value.}
}
\value{
Numeric vector.
}
\description{
The vector will accumulate each pair of adjacent elements, similar to \code{\link[base]{Reduce}} with the accumulate parameter.
If the summation of the two elements is greater than the threshold, the returned value of that cycle is zero. Primary use is within
\code{\link{assign_episode}} function to provide a speed boost in the looping procedure.
}
