% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate_age}
\alias{calculate_age}
\title{Calculate age}
\source{
\url{https://stackoverflow.com/questions/27096485/change-a-column-from-birth-date-to-age-in-r}
}
\usage{
calculate_age(dob, age_day = Sys.Date(), units = "years", floor = TRUE)
}
\arguments{
\item{dob}{Vector containing date of birth for age calculation.}

\item{age_day}{Reference date to calculate age as of that time-point.}

\item{units}{unit to measure age (default set to \code{"years"}). Passed to \code{\link[lubridate]{duration}}.}

\item{floor}{Boolean to determine to floor round the result (default set to \code{TRUE}).}
}
\value{
Age in \code{units}. Will be an integer if \code{floor = TRUE}.
}
\description{
\code{calculate_age} will determine age in years based upon two comparison dates.
}
\details{
Rounding is performed with \code{floor} so you are the same age up to the day before the next birthday
(e.g 5 years old from 5th birthday through the day before your 6th birthday). Set \code{floor = FALSE} to
return decimal ages, and change \code{units} for units other than years. Try combining with \code{\link{create_breaks}}
to make age groupings.
}
\note{
Adapted from SO post. Earlier methods used \code{as.integer((compare_date - dob) / 365.25)}
}
\examples{
\dontrun{
tempData <- data.frame(date_of_birth = as.Date(c('1991-01-01', '1990-02-04')), death_date = as.Date(c('1992-01-1', '2020-01-01')))
calculate_age(tempData$date_of_birth, tempData$death_date)
calculate_age(tempData$date_of_birth, tempData$death_date, units = 'minutes')
calculate_age(tempData$date_of_birth, tempData$death_date, floor = FALSE)
}
}
\author{
Gregor Thomas \url{https://stackoverflow.com/users/903061/gregor-thomas>}
}
