% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate_coalesce_origin}
\alias{calculate_coalesce_origin}
\title{Calculate originating column for a coalesce operation}
\usage{
calculate_coalesce_origin(data, cols)
}
\arguments{
\item{data}{Originating dataset that coalesce procedure occurred on.}

\item{cols}{Character vector of column names, in order coalesce would occur.}
}
\value{
Character vector with names of originating columns.
}
\description{
\code{calculate_coalesce_origin} will determine the column from which a value was filled during a coalesce procedure. Ensure
that all blanks or other undesired patterns are converted to \code{NA} before starting. Provide columns in order coalesce is expected to occur.
}
\examples{
testdata <- data.frame(col1 = c(1,2,3), col2 = c(NA, NA, 3), col3 = c(1, NA, 4))
calculate_coalesce_origin(testdata, c('col2', 'col3'))
}
