% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate_trimesters}
\alias{calculate_trimesters}
\title{Calculate trimesters and preterm births}
\usage{
calculate_trimesters(
  data,
  dob,
  gestation,
  avg_pregnancy = 40,
  avg_tri = 13,
  buffer = 0
)
}
\arguments{
\item{data}{A data object.}

\item{dob}{The column name for date of birth.}

\item{gestation}{The column name containing the numeric weeks of gestation (optional).}

\item{avg_pregnancy}{Integer value for the expected number of weeks in an average pregnancy (default: 40).}

\item{avg_tri}{Integer value for the expected length of each trimester (default: 13 weeks).}

\item{buffer}{Integer value. Will extend the date of first trimester. Useful when you want to keep average pregnancy the same but want to include a higher maximum of pregnancies (i.e. longer third trimester).}
}
\value{
A new dataset with the start and end dates of trimesters and whether or not the birth is suspected as pre-term.
}
\description{
\code{calculate_trimesters} calculates dates for pregnancy trimesters and if the birth was pre-term.
}
\details{
Uses the date of birth (required field) and any available information on the gestation time (in weeks)
to calculate the start and end of the trimesters, in separate columns, along with a flag for a suspected preterm birth.
If no gestation information is provided, the default average pregnancy length of 40 weeks is used. One can extend the length
of pregnancy by adjusting \code{avg_pregnancy} but it is advised to use \code{buffer} to extend the start date of the first trimester instead.
The average length of each trimester can also be adjusted from the default of 13 weeks; again, it is advised to leave this at the accepted default.
}
\examples{
tempData <- data.frame(date_of_birth = as.Date(c('1991-01-01', '1990-02-04', '1992-01-21')), gestation = c(35, 41, NA))
calculate_trimesters(tempData, date_of_birth, gestation)

}
