% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{connect_sas_libraries}
\alias{connect_sas_libraries}
\title{Create string vector for loading SAS libraries}
\usage{
connect_sas_libraries(lib_name = "temp", lib_path = "")
}
\arguments{
\item{lib_name}{A character vector for the SAS library name to be defined.}

\item{lib_path}{A character variable for the connection path to the library of interest.}
}
\value{
A character vector of SAS code for loading desired libraries.
}
\description{
\code{connect_sas_libraries} will create a character string of SAS code for loading the provided library and path information.
This is typically passed as an argument to \code{add_library} to \code{\link{load_from_sas}}. Assumes connecting to \emph{remote} libraries.
}
\examples{
\dontrun{
connect_sas_libraries(lib_name = 'mylib', lib_path = '/location/to/saslibrary/on/drive')
}
}
