% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert_blank2NA}
\alias{convert_blank2NA}
\title{Convert all columns with blank data to NA}
\usage{
convert_blank2NA(data, target = "", skip_dates = FALSE, modify_inplace = FALSE)
}
\arguments{
\item{data}{Data object.}

\item{target}{Replace all "" by default. Can replace with vector (e.g. \code{c('', ' ')})}

\item{skip_dates}{Logical; exclude date columns from operation.}

\item{modify_inplace}{Logical; uses data.table formats to reduce copies.}
}
\value{
Dataset with NA instead of blank ("") cells.
}
\description{
\code{convert_blank2NA} converts a column of with blank ("") to \code{NA}. Be certain to check column data types afterwards,
as dates and factors may be affected. Alternative: use related \code{tidyverse} verbs that can achieve a similar effect in
perhaps a few more steps. To improve the speed of conversion between dates and characters, a default format is provided. For large
data-sets, the data copy can be substantial and modifying the data in-place may be prefereable (uses \code{data.table} package)
}
\examples{
blankData <- data.frame(x = c(1, "", 2), y = c("", "Cool Stuff", "More Stuff"))
convert_blank2NA(blankData)

}
