% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert_elements}
\alias{convert_elements}
\title{Convert elements to a new value}
\usage{
convert_elements(data, col, target, replacement = NA, convert2factor = TRUE)
}
\arguments{
\item{data}{Data object.}

\item{col}{Character vector of column to conduct conversion.}

\item{target}{Character vector of levels for conversion.}

\item{replacement}{Value of replacement (typically a Character).}

\item{convert2factor}{Logical vector, if \code{TRUE}, will convert characters to factors.}
}
\value{
Dataset with reformatted columns.
}
\description{
\code{convert_elements} will convert a set of values in a column to a designated replacement value. This can be
viewed as a shortcut to using \code{\link[forcats]{fct_collapse}}.
}
