% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert_format}
\alias{convert_format}
\title{Convert column formats}
\usage{
convert_format(data, func, col_list, ...)
}
\arguments{
\item{data}{Data object.}

\item{func}{Function for converting data types.}

\item{col_list}{Character vector of column names.}

\item{...}{Additional parameters to pass to the function provided.}
}
\value{
Dataset with reformatted columns.
}
\description{
\code{convert_format} will accept a character vector of column names and convert them to a desired format using functions
such as \code{\link{as.factor}}.
}
\examples{
convert_format(iris, factor, c('Sepal.Width'))

}
