% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert_num2percent}
\alias{convert_num2percent}
\title{Convert number vector to percents}
\source{
Implemented from \href{https://stackoverflow.com/users/134830/richie-cotton}{Richie Cotton's} StackOverflow contribution:
\url{https://stackoverflow.com/questions/7145826/how-to-format-a-number-as-percentage-in-r}
}
\usage{
convert_num2percent(x, digits = 2, format = "f", suffix = "\%", ...)
}
\arguments{
\item{x}{Integer vector.}

\item{digits}{Integer for rounding precision.}

\item{format}{Character vector ('f' or 'g') passed to \code{\link{formatC}}.}

\item{suffix}{Character value to append to the end of the number (default is '\%').}

\item{...}{Additional parameters passed to \code{\link{formatC}}.}
}
\value{
Character vector.
}
\description{
\code{convert_num2percent} will convert all numbers into percent values with a defined precision.
As an alternative, \code{\link[scales]{percent}} or \code{\link[scales]{label_percent}} can be used.
}
\examples{
x <- c(-1, 0, 0.1, 0.5667, 1, 100.2)
convert_num2percent(x)

}
\seealso{
\code{\link[scales]{percent}}
\code{\link[scales]{label_percent}} (e.g. \code{label_percent()(x)})
}
\author{
Richie Cotton.
}
