% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert_table2png}
\alias{convert_table2png}
\title{Convert a table object to a png file}
\source{
Inspired/adapted by a post from Michael Yan's blog: \url{http://thinkdatascience.com/}
}
\usage{
convert_table2png(obj, name, rez = 600, interpreter = "xtable", ...)
}
\arguments{
\item{obj}{A table object (huxtable, xtable, or tables).}

\item{name}{Character vector for naming output.}

\item{rez}{Integer value for desired resolution (default is 600).}

\item{interpreter}{Character vector to define which type of table object was used (huxtable, xtable, or tables).}

\item{...}{Additional parameters passed to \code{\link[tools]{texi2dvi}}.}
}
\value{
A png file saved to working directory.
}
\description{
\code{convert_table2png} will take various types of table object formats and convert them to a png image.
}
\details{
To create the png file, a LaTeX distribution, and texi2dvi (from base R) are required. It is recommended to use
\code{\link[tinytex]{install_tinytex}}. Various LaTeX packages may be needed, specifically 'tabularx' and 'dvipng' should be
installed using \code{tinytex::tlmgr_install('dvipng')}

On some Linux machines, the texi2dvi operation may not work without additional configuration. As such, the parameter \code{texi2dvi} parameter
should be set to \code{emulation}. This has been tested with \code{\link[xtable]{xtable}} but the other interpreters may have additional
dependencies beyond this work-around.
}
\examples{
convert_table2png(mtcars, 'carsTable', 600, interpreter = 'xtable')
convert_table2png(mtcars, 'carsTable', 600, interpreter = 'xtable', texi2dvi = 'emulation')

}
