% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert_wk_flu2calendar}
\alias{convert_wk_flu2calendar}
\title{Convert flu week to calendar week}
\usage{
convert_wk_flu2calendar(week, flu_wk_start = 34)
}
\arguments{
\item{week}{An integer between 1 and 52 representing the flu week to convert.}

\item{flu_wk_start}{Week of the year that flu season begins; default set to 34.}
}
\value{
An integer vector representing the calendar week.
}
\description{
\code{convert_wk_flu2calendar} will convert seasonal flu weeks to calendar weeks.
}
\details{
Converts between seasonal flu weeks and calendar weeks, which is often used in influenza reporting.
Calendar weeks starts from week 1 in January to week 52/53. Seasonal flu weeks begins at the start of the
influenza season (Week 1) up to the week prior to the following influenza season (week 52/53). The start point of
the influenza season can change from year to year. You can perform the opposite operation via \code{\link{convert_wk_calendar2flu}}.
}
\note{
Adapted from original with courtesy of M. Parsons.
}
\examples{
convert_wk_flu2calendar(18)
convert_wk_flu2calendar(19)
convert_wk_flu2calendar(c(35, 32, 1, 23))
convert_wk_flu2calendar(18, 45)

}
\seealso{
\code{\link{convert_wk_calendar2flu}}
}
