% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{count_decimals}
\alias{count_decimals}
\title{Count decimal places}
\usage{
count_decimals(x)
}
\arguments{
\item{x}{a vector of numeric values.}
}
\value{
Length of each values decimal places.
}
\description{
Examine an number and return the number of decimals present.
}
\details{
Uses conversions from a numeric to character to splice off and count how many decimals are provided.
Supports vectorization.
}
\examples{
testing <- c(0, 100, 100.1, 101.10101, 100.00)
count_decimals(testing)
}
