% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_circle}
\alias{create_circle}
\title{Create x,y coordinates of a circle}
\source{
Adapted from \href{https://stackoverflow.com/users/1457051/hrbrmstr}{Bob Rudis'} StackOverflow post:
\url{https://stackoverflow.com/questions/31635732/stacked-bubble-chart-bottom-aligned}
}
\description{
\code{createCircle} will create the x,y coordinates for a circle to be made using the polygon geom in \code{ggplot2}. This code is
internal to \code{\link{StatBullet}}.
}
\author{
Bob Rudis
}
\keyword{internal}
