% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_datalist}
\alias{create_datalist}
\title{Create a data list}
\usage{
create_datalist(data_names, envir = .GlobalEnv)
}
\arguments{
\item{data_names}{A character vector of all datasets in the selected environment to collapse into a named list.}

\item{envir}{An environment object, defaulted to \code{.GlobalEnv}}
}
\value{
A named list.
}
\description{
\code{create_datalist} creates data list by using data names and combining them all into a named list.
}
\examples{
dataset1 <- mtcars
dataset2 <- iris
combined_datalist <- create_datalist(c('dataset1', 'dataset2'))

}
