% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_hypercube}
\alias{create_hypercube}
\title{Create a data hypercube}
\usage{
create_hypercube(
  data,
  columns,
  drop_sum_columns = "",
  use_na = "ifany",
  drop_na_levels = TRUE,
  label_for_sum = NULL
)
}
\arguments{
\item{data}{Data object.}

\item{columns}{Character vector, must be more than 1 column provided.}

\item{drop_sum_columns}{Character vector of which columns to drop the 'sum' levels.}

\item{use_na}{Character vector regarding if \code{NA} levels should be included in tallies, passed to \code{table}.}

\item{drop_na_levels}{Logical vector, if \code{TRUE} all \code{NA} factor levels are dropped.}

\item{label_for_sum}{Character vector, a label to replace the default text of 'sum' for column totals.}
}
\value{
A \code{data.frame} or \code{tibble} is returned
}
\description{
\code{create_hypercube} creates a data cube from a flat data table. This is very similar to \code{groupingsets()} in the \code{data.table} package.
}
\details{
If \code{drop_na_levels} is set to \code{TRUE} all factor levels which are \code{NA} will be removed. A label can also be set for all the 'sum' levels;
typically these 'sum' levels are \code{NA_character_} but they will not be removed by \code{drop_na_levels} (only NA levels that exist in the data prior to the summarizing
are dropped.
}
\examples{
create_hypercube(mtcars, columns = c('gear', 'cyl'), drop_sum_columns = 'cyl')

}
