% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_pretty_int_brks}
\alias{create_pretty_int_brks}
\title{Create clean integer breaks for scales}
\usage{
create_pretty_int_brks(x, max_multiplier = 1.03)
}
\arguments{
\item{x}{Data vector.}

\item{max_multiplier}{Integer vector to control the scaling of the breaks.}
}
\description{
\code{create_pretty_int_brks} will take an input of values and generate clean integer break points,
scaled by a multiplier variable.
}
\examples{
data_for_breaks <- c(1,2,3.8,5,25.5,60)
create_pretty_int_brks(data_for_breaks, max_multiplier = 1.05)

}
