% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_randomstring}
\alias{create_randomstring}
\title{Create a random string of n length.}
\usage{
create_randomstring(length = 8L)
}
\arguments{
\item{length}{Numeric value, will be coerced into an integer. Defaults to \code{8L}.}
}
\value{
Character vector matching provided \code{length}.
}
\description{
Using \code{\link{paste0}} and \code{\link{sample}}, a specified length of randomly selected letters and numbers
will be pasted together.
}
\examples{
create_randomstring()
create_randomstring(5L)
create_randomstring(17)

}
