% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{crude_rate}
\alias{crude_rate}
\title{Calculate crude rates (internal)}
\usage{
crude_rate(num, denom, per)
}
\arguments{
\item{num}{Numerator value.}

\item{denom}{Denominator value}

\item{per}{Numeric value for rate scale (i.e. per X population).}
}
\description{
Calculate crude rates (internal)
}
