% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_file_content}
\alias{find_file_content}
\title{Find files containing pattern match}
\usage{
find_file_content(
  path,
  file_pattern = NULL,
  text_pattern,
  recursive = FALSE,
  locale = readr::default_locale(),
  ...
)
}
\arguments{
\item{path}{File path that contains files of interest (can look into child folders using \code{recursive = T})}

\item{file_pattern}{Basic regular expression to pick up file name pattern (e.g. \code{'\\.(R|Rmd)'})}

\item{text_pattern}{Grab the n-th file from the list (descending order by date). Default set to \code{NULL}, entire list is returned. Set to \code{1} if want just the latest match returned.}

\item{recursive}{Reduce set of discovered file based upon a date threshold (default: \code{NULL}, no filtering occurs).}

\item{locale}{Passed to \code{\link[readr]{read_file}}, controls encoding and similar file properties.}

\item{...}{Additional parameters passed to \code{\link{grepl}}, helpful to add \code{perl} capabilities in pattern search.}
}
\value{
Returns character vector for file names with matches
}
\description{
Identify files within provided path that contain text that match a particular pattern.
}
\details{
This function is often useful to quickly scan folders containing code that you need to find a particular key-phrase.
}
\examples{

# Find pattern within set of files
find_file_content('./R', file_pattern = '\\\\.(R|Rmd)', text_pattern = 'sapply')

}
