% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_victoriaday}
\alias{find_victoriaday}
\title{Find Victoria day by year}
\usage{
find_victoriaday(year)
}
\arguments{
\item{year}{character or numeric for 4 digit year.}
}
\value{
Date object in YYYY-MM-DD format.
}
\description{
Determine the day of the year Victoria day (Monday before 25th of May) is observed as a holiday.
}
\examples{
find_victoriaday(2021)
lapply(c(1990:2000), find_victoriaday)
}
