% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{geom_bullseye}
\alias{geom_bullseye}
\title{Create bottom-aligned bullet chart}
\usage{
geom_bullseye(
  mapping = NULL,
  data = NULL,
  geom = "polygon",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}} or
\code{\link[ggplot2:aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}
}
\description{
\code{geom_bullseye} create a bullet chart with all the circles aligned to the bottom.
}
\details{
Since this is a specialized plot, it requires data in a specific format (similar to line charts). The \code{x} parameter
must be the proportion of each category (i.e. the percentage one category is from the total). A \code{group} must also be
provided in a form of \code{fill, group, color}. It is also recommended to add \code{\link[ggplot2]{coord_equal}} and to adjust
the theme to something more minimal (e.g. \code{ggthemes::theme_map()}). Labeling the circles may be difficult at this time, but
a future version may make this easier without having to do additional calculations for ideal placement. Factor levels may
also need to be reordered (e.g. using \code{forcats}). Alternatively, instead of using this custom Stat, one could just use programming methods with ggplot2.
}
\examples{
\dontrun{
library(ggplot2)
library(magrittr)
library(tibble)
library(forcats)

# Create data
data <- tribble(~Grouping, ~Cases,
                "Step1", 4000,
                "Step2", 2300,
                "ThirdStep", 200,
                "Step4", 20)

# Plot data
ggplot(data = data, aes(x = Cases, fill = fct_rev(fct_reorder(Grouping, Cases)))) +
  geom_bullseye()
  coord_equal() +
  ggthemes::theme_map() +
  theme(legend.title = element.blank(),
        legend.position = 'right')
 }

}
