% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\docType{data}
\name{GeomXSpline}
\alias{GeomXSpline}
\title{ggproto of geom_xspline for ggplot2}
\description{
\code{GeomXSpline} is the \code{ggproto} object required for \code{\link{geom_xspline}}.
}
\keyword{internal}
