% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{label_notation_brackets}
\alias{label_notation_brackets}
\title{Create middle labels with 'bracket format' (internal)}
\usage{
label_notation_brackets(
  breaks,
  brk_length,
  divider,
  left.open,
  rightmost.closed
)
}
\arguments{
\item{breaks}{Numeric vector of breakpoints supplied to \code{\link{findInterval}}.}

\item{brk_length}{Length of number of breaks provided to parent function.}

\item{divider}{Character defining the symbol separating adjacent label values (default is a dash).}

\item{rightmost.closed}{Logical; passed to \code{\link{findInterval}}.}
}
\description{
Create middle labels with 'bracket format' (internal)
}
\seealso{
\code{\link{create_breaks}}
}
