% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{label_notation_precision}
\alias{label_notation_precision}
\title{Create middle labels with 'precision format' (internal)}
\usage{
label_notation_precision(
  breaks,
  brk_length,
  precision,
  divider,
  left.open,
  rightmost.closed
)
}
\arguments{
\item{breaks}{Numeric vector of breakpoints supplied to \code{\link{findInterval}}.}

\item{brk_length}{Length of number of breaks provided to parent function.}

\item{precision}{Numeric value, determines how to adjust boundary of labels. Default will determine from provided breaks (e.g. 1 decimal place means precision of 0.1).}

\item{divider}{Character defining the symbol separating adjacent label values (default is a dash).}

\item{rightmost.closed}{Logical; passed to \code{\link{findInterval}}.}
}
\description{
Create middle labels with 'precision format' (internal)
}
\seealso{
\code{\link{create_breaks}}
}
