% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_weekdays}
\alias{list_weekdays}
\title{Find the all dates for weekday in a month}
\usage{
list_weekdays(year, month, weekday)
}
\arguments{
\item{year}{character or numeric for 4 digit year.}

\item{month}{character or numeric (01-12) for month.}

\item{weekday}{character value for day of the week (e.g. 'Monday').}
}
\value{
Date vector object in YYYY-MM-DD format.
}
\description{
Determine the exact days (i.e. YYYY-MM-DD) that a specific day of the week falls on.
}
\examples{
list_weekdays(2021, 12, 'Monday')
list_weekdays(2021, 12, 'Monday')[1]
list_weekdays('2021', 12, 'Monday')
lapply(c(1990:2000), list_weekdays, month = 12, weekday = 'Monday')
}
