% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{load_excel_sheets}
\alias{load_excel_sheets}
\title{Load all sheets from an excel file into R}
\source{
Adapted from \href{https://stackoverflow.com/users/180892/jeromy-anglim}{Jeromy Anglim's} stackoverflow contribution:
\url{https://stackoverflow.com/questions/12945687/read-all-worksheets-in-an-excel-workbook-into-an-r-list-with-data-frames}
}
\usage{
load_excel_sheets(filename, tibble = FALSE, ...)
}
\arguments{
\item{filename}{A character string for the file path where the excel file is located.}

\item{tibble}{A logical value, if \code{TRUE} will export all sheets as a \code{tibble} object.}

\item{...}{Additional parameters passed to \code{read_excel}.}
}
\value{
Datasets for all sheets within excel file.
}
\description{
\code{load_excel_sheets} will load all sheets from an excel file.
}
\note{
Requires the \code{readxl} package to operate.
}
\examples{
\dontrun{
loaded_sheets <- load_excel_sheets("myexcelfile.xls")
}
}
