% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{maxN}
\alias{maxN}
\alias{minN}
\title{Determine the top max or min values}
\source{
Adapted from \href{https://stackoverflow.com/users/345660/zach}{Zachary Mayer's} stackoverflow contribution:
\url{https://stackoverflow.com/questions/2453326/fastest-way-to-find-second-third-highest-lowest-value-in-vector-or-column}
}
\usage{
maxN(x, N = 1)

minN(x, N = 1)
}
\arguments{
\item{x}{A data object.}

\item{N}{Character vector of the column name.}
}
\value{
The Nth max or min value (numeric).
}
\description{
Determine the top maximum or minimum values in a vector.
}
\examples{
tempData <- c(1,4,1,7,2)
maxN(tempData, N=2)
minN(tempData, N=2)

}
