% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{pass_code_to_sas}
\alias{pass_code_to_sas}
\title{Pass code from R to SAS}
\usage{
pass_code_to_sas(
  sas_file_list = NULL,
  inputstring = NULL,
  sas_path = "C:/Program Files/SASHome/SASFoundation/9.4/sas.exe",
  config_file = NULL
)
}
\arguments{
\item{sas_file_list}{A list of strings defining the location and name of SAS
code files.}

\item{inputstring}{A string vector of SAS code.}

\item{sas_path}{File path to SAS 9.4 executable; default set to a rough standard of: \code{C:/Program Files/SASHome/SASFoundation/9.4/sas.exe}.}

\item{config_file}{File path for SAS 9.4 configuration file (e.g. C:/.../SAS9.4_Prod/SASV9.CFG).}
}
\description{
\code{pass_code_to_sas} will pass SAS code (typed in R and/or list of SAS
files) from R to an installed local version of SAS.
}
\examples{
\dontrun{
sas_code_string <- 'DATA TEMP; do i = 1 to 10; age = i+1; output; end; run;'
pass_code_to_sas(inputstring = sas_code_string, sas_path = "C:/Program Files/SASHome/SASFoundation/9.4/sas.exe")
}
}
