% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{read_sqlite}
\alias{read_sqlite}
\title{Quickly query a SQLite database}
\usage{
read_sqlite(db_name, query, extended_types = TRUE, ...)
}
\arguments{
\item{db_name}{Database name (full or relative path must be part of the name).}

\item{query}{Character vector containing the SQL query.}

\item{extended_types}{RSQLite parameter for additional data types such as dates (default: \code{TRUE})}

\item{...}{Additional parameters to be passed to \code{\link[DBI]{dbGetQuery}}.}
}
\description{
Access a SQLite file and pass a standard query. This function deals with all the {DBI} work and will open/close the connection
automatically. If you need to perform multiple operations it is advised to work with {DBI} directly.
}
\examples{
\dontrun{
# Create a new DB and add first table in one swoop
write2sqlite('mydb.sqlite', mtcars, 'firsttable')

read_sqlite('mydb.sqlite', "SELECT * FROM firsttable")
read_sqlite('mydb.sqlite', "SELECT gear, count(*) AS N FROM firsttable GROUP BY gear")
}
}
\seealso{
write2sqlite
}
