% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rename_file}
\alias{rename_file}
\title{Rename and relocate a file}
\usage{
rename_file(from, to)
}
\arguments{
\item{from}{File path and name of original file.}

\item{to}{File path and name of altered/moved file.}
}
\description{
\code{rename_file} will rename and/or move a file; this is useful for adjusting R markdown output locations.
}
