% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reorder_within}
\alias{reorder_within}
\alias{scale_x_reordered}
\alias{scale_y_reordered}
\title{Reorder the levels within groups}
\source{
David Robinson's personal package, specifically "reoder_within.R"
\url{https://github.com/dgrtwo/drlib/blob/master/R/reorder_within.R}
}
\usage{
reorder_within(x, by, within, fun = mean, sep = "___", ...)

scale_x_reordered(..., sep = "___")

scale_y_reordered(..., sep = "___")
}
\arguments{
\item{x}{Vector to reorder.}

\item{by}{Vector, of same length as \code{x}, used as reference for reordering.}

\item{within}{Vector, of same length as \code{x}, used for facetting.}

\item{fun}{Function used within subsets to determine the ordering result (\code{mean} by default).}

\item{sep}{Character used to separate labels for the reordering.}

\item{...}{Additional arguments passed to \code{\link{reorder}} for \code{reorder_within}. For scale functions,
additional arguments passed to either \code{\link[ggplot2]{scale_y_discrete}} or \code{\link[ggplot2]{scale_x_discrete}}.}
}
\description{
A set of functions to reorder factors within faceted groups, useful for plotting bar charts with \code{ggplot2}.
}
\details{
First use \code{reorder_within} during \code{aes()} step for the factor level and the variable used for faceting.
Then, use the scale functions (\code{scale_xreordered}, \code{scale_y_reordered}) to adjust the scale to re-order.
}
\examples{
library(ggplot2)
library(dplyr)
library(tidyr)

iris_gathered <- gather(iris, metric, value, -Species)

ggplot(iris_gathered, aes(reorder_within(Species, value, metric), value)) +
  geom_bar(stat = "identity") +
  scale_x_reordered() +
  facet_wrap(~ metric, scales = "free_x")
}
\seealso{
"Ordering categories within gplot2 Facets" by Tyler Rinker:
\url{https://trinkerrstuff.wordpress.com/2016/12/23/ordering-categories-within-ggplot2-facets/}
}
\author{
David Robinson.
}
