% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{repath}
\alias{repath}
\title{Convert Windows-style paths to UNIX-style}
\source{
Adapted from \href{https://stackoverflow.com/users/378085/tom}{Tom's} stackoverflow contribution:
\url{https://stackoverflow.com/questions/1189759/expert-r-users-whats-in-your-rprofile/12703931#12703931}
}
\usage{
repath()
}
\description{
\code{repath} will convert a Windows-style file path to UNIX-style, which can then be used in R code. The function
is interactive and should be called without any arguments provided. A prompt will appear for input and the converted
file path will be copied to the clipboard.
}
\examples{
repath()

}
