% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{round_cols}
\alias{round_cols}
\title{Round values in select columns}
\usage{
round_cols(data, col_list, dec = 2)
}
\arguments{
\item{data}{A data object.}

\item{col_list}{A character vector of specific column names to round.}

\item{dec}{A numeric value to control precision of rounding.}
}
\value{
A new dataset object with select columns rounded.
}
\description{
\code{round_cols} will round values in a pre-defined selection of columns in a dataset.
}
\details{
The function is vectorized so that multiple inputs can be examined without having to define any specific loops.
}
\examples{
tempData <- data.frame(rate = as.numeric(c(10.344, 0.359 , 1023.49)), value = as.numeric(c(10, 1000, 3200.2)))
round_cols(tempData, c('rate'), 2)

}
