% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\docType{data}
\name{StatBullseye}
\alias{StatBullseye}
\title{\code{ggproto} for \code{geom_bullseye}}
\source{
Adapted from \href{https://stackoverflow.com/users/1457051/hrbrmstr}{Bob Rudis'} StackOverflow post:
\url{https://stackoverflow.com/questions/31635732/stacked-bubble-chart-bottom-aligned}
}
\description{
\code{StatBullseye} is the \code{ggproto} object required for \code{\link{geom_bullseye}}.
}
\keyword{internal}
