% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{std_err}
\alias{std_err}
\title{Calculate standard error (internal)}
\usage{
std_err(num, denom, per, modified_wald = FALSE)
}
\arguments{
\item{num}{Numerator value.}

\item{denom}{Denominator value}

\item{per}{Numeric value for rate scale (i.e. per X population).}

\item{modified_wald}{Boolean for if the standard error should be calculated with modified Wald interval.}
}
\description{
Calculate standard error (internal)
}
