% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{transfer_curl}
\alias{transfer_curl}
\title{Transfer content between file systems via cURL}
\usage{
transfer_curl(
  file = NULL,
  direction = NULL,
  connection = NULL,
  handle_options = list(),
  drop_location = NULL
)
}
\arguments{
\item{file}{Character string. When uploading, should be the file path and name, when downloading it should just be the file name and extension.}

\item{direction}{Character string, either 'upload' or 'download'.}

\item{connection}{Character string that is used to connect to the desired file system.}

\item{handle_options}{List of names elements for the cURL handle (for sftp connections, typically username and password are provided).}

\item{drop_location}{Character string that provides the path to place the downloaded file from winSCP.}
}
\value{
Dataset will be saved in the file system.
}
\description{
\code{transfer_curl} will connect to the (secure) file transfer protocol available with cURL and transfer (upload or download)
a specific file of interest.
}
\details{
\code{transfer_curl} is a thin wrapper around \code{curl} functionality; more specifically, the core process uses \code{\link[curl]{curl_download}} and
\code{\link[curl]{curl_upload}}. cURL requires a 'handle' to be specified, which as the name suggests, specifies how cURL will 'handle' the server
request. Details such as username, password, verbose can be specified in the handle. See the examples for some common patterns. A possible limitation is
the transfer speed of this process. The default transfer speed as controlled through \href{https://curl.se/libcurl/c/curl_easy_setopt.html}{URL options}; if you have
over several MBs to transfer, try adjusting the \code{buffersize}, \code{upload_buffersize}, \code{max_recv_speed_large}, and \code{max_send_speed_large} options passed
to the \code{handle_options} parameter. Use \code{\link[curl]{curl_options}} to see what options are available, if some are expected but not found, consider upgrading cURL.
}
\seealso{

}
