% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{write2sqlite}
\alias{write2sqlite}
\title{Quickly create and write to a SQLite database}
\usage{
write2sqlite(
  db_name,
  data,
  tbl_name,
  extended_types = TRUE,
  date2text = TRUE,
  logging = FALSE,
  ...
)
}
\arguments{
\item{db_name}{Database name (either path to existing, or name of new database (e.g. name.sqlite, name.db)).}

\item{data}{Dataset to add to database.}

\item{tbl_name}{Name to assign to table associated with the provided data.}

\item{extended_types}{RSQLite parameter for additional data types such as dates (default: \code{TRUE}),}

\item{date2text}{Detect which columns are dates and parse them to characters (instead of numeric) to store in SQLite (default: \code{TRUE}),}

\item{logging}{Boolean, determines if logging table named 'log_book' is used in the writing operation (default: \code{FALSE}),}

\item{...}{Additional parameters to be passed to \code{\link[DBI]{dbWriteTable}}.}
}
\description{
\code{write2sqlite} is a light wrapper function around various \code{RSQLite} and \code{DBI} functions to make it easier to
quickly create a database, tables, and append to existing tables.
}
\details{
This function requires some of the more recent features of SQLite, including 'extended types' and date formatting capabilities.
To ensure date formatting is preserved, \code{write2sqlite} checks column types for the table being added. Using this function also
controls the connection open/closing automatically (given the provided DB name).

An optional logging feature is also available. This creates a column called 'log_book' and will track when a table was updated.
}
\examples{
\dontrun{
# Use an existing dataset to add data
write2sqlite('path/to/my/dbfile.sqlite', mtcars, 'boringcars')

# Replace an existing dataset
newmtcars <- mtcars[1:10]
write2sqlite('path/to/my/dbfile.sqlite', newmtcars, 'boringcars', overwrite = TRUE)

# Create a new DB and add first table in one swoop
write2sqlite('mydb.sqlite', newmtcars, 'firsttable')
}
}
\seealso{
read_sqlite
}
