% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{write_sql_log}
\alias{write_sql_log}
\title{Write to logging column in SQLite}
\usage{
write_sql_log(db_name, message = NA_character_, extended_types = TRUE, ...)
}
\arguments{
\item{db_name}{Database name (full or relative path must be part of the name).}

\item{message}{Logging message to write to log_book table.}

\item{extended_types}{RSQLite parameter for additional data types such as dates (default: \code{TRUE})}

\item{...}{Additional parameters to be passed to \code{\link[DBI]{dbWriteTable}}.}
}
\description{
Used internally by write2sqlite
}
