% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert_date2fluseason}
\alias{convert_date2fluseason}
\title{Convert calendar date to flu season}
\usage{
convert_date2fluseason(
  date,
  format = "\%Y-\%m-\%d",
  flu_wk_start = 34,
  return_values
)
}
\arguments{
\item{date}{Character vector in date format of \code{'YYYY-mm-dd'}.}

\item{format}{Character vector following \code{\link[base]{strptime}}; defaults to \code{"\%Y-\%m-\%d"}.}

\item{flu_wk_start}{Week of the year that flu season begins, all entries prior to that week will be in prior season; default set to 34.}

\item{return_values}{Character vector of which values to return, default is set to all ('week', 'month', 'year', 'season').}
}
\value{
List containing vectors of week, month, year, and season related to provided dates.
}
\description{
Convert a vector of input dates to flu season details. Output on week, month, year, and season are output by default.
}
\details{
The primary purpose of this function is to reassign week 53 dates into week 1 and week 52. It will also provide flu season
assignment based upon a flu week start date as reference. To improve speed, indexing methods are used to assign values instead of
\code{\link{ifelse}} (increased speed by over 5-fold).
}
\note{
Adapted from original with courtesy of M. Ware.
}
\examples{
date_list <- c('2020-08-30', '2020-09-01', '2020-09-23', '2020-01-01', '2019-12-31', '2018-01-01', '2017-01-01', '2016-01-01')
convert_date2fluseason(date_list)
convert_date2fluseason(date_list, return_values = 'season')
convert_date2fluseason(date_list, flu_wk_start = 40)

}
\seealso{
\code{\link{convert_wk_flu2calendar}}, \code{\link{convert_wk_calendar2flu}}
}
